cmake: avoid use of system-provided <zzip/_config.h> header
Our dependents on the zzip library will fail to compile in case where
the host system does not have zziplib installed as the compiler cannot
find the <zzip/_config.h> header. Maybe even more importantly, we were
always using the system-provided "_config.h" instead of the one that we
were generating.

This is due to us generating the "_config.h" file into the binary
directory of the "zzip" subdirectory. But as we did not always add the
top-level CMAKE_BINARY_DIR to each of the dependents' include search
paths, those would pick either the system-provided header file or fail
to compile.

Fix this by properly setting up the include directories, adding
CMAKE_BINARY_DIR to all of our dependent targets.