cmake: define ZZIP_PACKAGE and ZZIP_VERSION
Many binaries rely on defines for ZZIP_PACKAGE and ZZIP_VERSION to be
available in order to display version information on the command line.
While the old autotools-based build system correctly set up those
defines in the corresponding config header, the CMake does not do this
correctly. There are indeed `#cmakedefine`s  in the "_config.h.cmake"
file, but we never set those variables in the corresponding
"CMakeLists.txt" file.

This problem only manifests itself if there is no <zzip/_config.h> file
installed on the host system, as the binaries' include paths are not
corrrectly set up and thus we never find the "_config.h" file generated
by CMake. This is being fixed in a later patch, though.

Fix part of the issue by setting ZZIP_PACKAGE and ZZIP_VERSION to the
project's name and version, respectively.